Read fields even for empty messages
When a message definition doesnt have any fields, it could be that in the future
the definition has fields. The message on the wire could have unknown fields,
even if our definition does not have any fields. So, we should read fields
for "empty" messages and store the unknown fields so that we can pass them
through if needed.